llm_fundamentals/rag-customer-support-chatbot/requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.38
chromadb>=0.4.0
streamlit>=1.28.0
openai>=1.0.0
pypdf>=3.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0